+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -1,5 +1,9 @@
 # GreenEnergyHub.Charges.Libraries Release notes
 
+## Version 1.0.10
+
+Integration tests re-enabled in Energinet GitHub pipeline.
+
 ## Version 1.0.9
 
 Protobuf contracts common for both Energinet.Charges.Libraries and elsewhere in the Charges domain are now located in one common place to avoid duplication.